// frontend/src/lib/sdk/models/UserCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserCreate = {
    email: string;
    phone?: (string | null);
    password: string;
    first_name: string;
    last_name: string;
    status?: string;
};