// frontend/src/lib/sdk/models/UserUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserUpdate = {
    email?: (string | null);
    phone?: (string | null);
    password?: (string | null);
    first_name?: (string | null);
    last_name?: (string | null);
    status?: (string | null);
};