// frontend/src/lib/sdk/models/StoreResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StoreResponse = {
    id: number;
    company_id: number;
    company_gstin_id: (number | null);
    name: string;
    address: string;
    is_franchise: boolean;
    status: string;
    timezone: string;
    invoice_series_prefix: string;
    created_at: string;
    updated_at: string;
};